fix: break component into subcomponents
