+++ packages/odyssey-react-mui/src/labs/Stepper/Stepper.tsx
@@ -17,118 +17,20 @@ import {
   Stepper as MuiStepper,
 } from "@mui/material";
 import { memo, useCallback, useMemo } from "react";
-import { useTranslation } from "react-i18next";
-import { Button } from "../../Buttons";
-import { HtmlProps } from "../../HtmlProps";
 import { CheckIcon } from "../../icons.generated";
 import {
   DesignTokens,
   useOdysseyDesignTokens,
 } from "../../OdysseyDesignTokensContext";
-
-export type StepData = {
-  /**
-   * The label text for the step
-   */
-  label: string;
-  /**
-   * Optional description text below step label
-   */
-  description?: string;
-};
-
-export type StepperProps = {
-  /**
-   * Aria label for the stepper container
-   */
-  ariaLabel?: string;
-  /**
-   * Current active step (0-based index)
-   */
-  activeStep: number;
-  /**
-   * Allow navigation to completed steps
-   */
-  allowBackStep?: boolean;
-  /**
-   * Allow skipping to future steps
-   */
-  nonLinear?: boolean;
-  /**
-   * Layout orientation of the stepper
-   */
-  orientation?: "horizontal" | "vertical";
-  /**
-   * Visual variant of steps
-   */
-  variant?: "numeric" | "nonNumeric";
-  /**
-   * Array of step data
-   */
-  steps: StepData[];
-  /**
-   * Callback fired when a step is clicked
-   */
-  onChange?: (step: number) => void;
-  /**
-   * Button label for the previous navigation button
-   */
-  previousButtonLabel?: string;
-  /**
-   * Button label for the next navigation button
-   */
-  nextButtonLabel?: string;
-} & Pick<HtmlProps, "testId" | "translate">;
-
-const createShouldForwardProp = (excludedProps: string[]) => (prop: string) =>
-  !excludedProps.includes(prop);
-
-const shouldForwardStepProps = createShouldForwardProp([
-  "odysseyDesignTokens",
-  "orientation",
-  "isClickable",
-]);
-
-const shouldForwardStepIconContainerProps = createShouldForwardProp([
-  "completed",
-  "active",
-  "variant",
-  "odysseyDesignTokens",
-  "nonLinear",
-]);
-
-const shouldForwardStepperProps = createShouldForwardProp([
-  "odysseyDesignTokens",
-  "allowBackStep",
-  "nonLinear",
-  "stepVariant",
-]);
-
-const shouldForwardStepDescriptionProps = createShouldForwardProp([
-  "odysseyDesignTokens",
-  "completed",
-  "active",
-  "orientation",
-]);
-
-const shouldForwardStepNumberProps = createShouldForwardProp([
-  "odysseyDesignTokens",
-  "completed",
-  "active",
-  "nonLinear",
-]);
-
-const shouldForwardStepperDotProps = createShouldForwardProp([
-  "status",
-  "odysseyDesignTokens",
-  "isClickable",
-]);
-
-const shouldForwardStepperNavigationProps = createShouldForwardProp([
-  "odysseyDesignTokens",
-]);
-
-const shouldForwardNavigationSectionProps = createShouldForwardProp(["align"]);
+import { StepperProps } from "./Stepper.types";
+import {
+  shouldForwardStepDescriptionProps,
+  shouldForwardStepIconContainerProps,
+  shouldForwardStepNumberProps,
+  shouldForwardStepperProps,
+  shouldForwardStepProps,
+} from "./Stepper.utils";
+//import { useTranslation } from "react-i18next";
 
 const StyledStep = styled(MuiStep, {
   shouldForwardProp: shouldForwardStepProps,
@@ -151,11 +53,11 @@ const StyledStep = styled(MuiStep, {
 const StepperContainer = styled(MuiStepper, {
   shouldForwardProp: shouldForwardStepperProps,
 })<{
+  nonLinear?: boolean;
   odysseyDesignTokens: DesignTokens;
   orientation?: "horizontal" | "vertical";
-  nonLinear?: boolean;
   stepVariant?: "numeric" | "nonNumeric";
-}>(({ orientation, odysseyDesignTokens, nonLinear, stepVariant }) => {
+}>(({ nonLinear, odysseyDesignTokens, orientation, stepVariant }) => {
   return {
     alignItems: "start",
     ...(orientation === "horizontal" && {
@@ -251,12 +153,12 @@ const StepperContainer = styled(MuiStepper, {
 const StyledStepIconContainer = styled("div", {
   shouldForwardProp: shouldForwardStepIconContainerProps,
 })<{
-  completed: boolean;
   active: boolean;
-  variant: "numeric" | "nonNumeric";
+  completed: boolean;
   nonLinear: boolean;
   odysseyDesignTokens: DesignTokens;
-}>(({ completed, active, variant, nonLinear, odysseyDesignTokens }) => ({
+  variant: "numeric" | "nonNumeric";
+}>(({ active, completed, variant, nonLinear, odysseyDesignTokens }) => ({
   width:
     variant === "numeric"
       ? odysseyDesignTokens.Spacing5
@@ -308,28 +210,28 @@ const StyledStepIconContainer = styled("div", {
 const StepLabel = styled(MuiStepLabel, {
   shouldForwardProp: (prop) =>
     ![
-      "odysseyDesignTokens",
-      "completed",
       "active",
       "allowBackStep",
+      "completed",
       "nonLinear",
+      "odysseyDesignTokens",
       "orientation",
       "variant",
     ].includes(prop as string),
 })<{
-  odysseyDesignTokens: DesignTokens;
-  completed: boolean;
   active: boolean;
   allowBackStep?: boolean;
+  completed: boolean;
+  odysseyDesignTokens: DesignTokens;
   nonLinear?: boolean;
   orientation?: "horizontal" | "vertical";
   variant?: "numeric" | "nonNumeric";
 }>(
   ({
-    completed,
     active,
-    nonLinear,
     allowBackStep,
+    completed,
+    nonLinear,
     odysseyDesignTokens,
     orientation,
     variant,
@@ -375,11 +277,11 @@ const StepLabel = styled(MuiStepLabel, {
 const StyledStepDescription = styled("div", {
   shouldForwardProp: shouldForwardStepDescriptionProps,
 })<{
-  odysseyDesignTokens: DesignTokens;
-  completed: boolean;
   active: boolean;
+  completed: boolean;
+  odysseyDesignTokens: DesignTokens;
   orientation?: "horizontal" | "vertical";
-}>(({ completed, active, odysseyDesignTokens, orientation }) => ({
+}>(({ active, completed, odysseyDesignTokens, orientation }) => ({
   fontSize: odysseyDesignTokens.TypographySizeSubordinate,
   fontWeight: odysseyDesignTokens.TypographyWeightBody,
   lineHeight: odysseyDesignTokens.TypographyLineHeightBody,
@@ -413,226 +315,6 @@ const StyledStepNumber = styled("span", {
   },
 }));
 
-const StyledStepperDot = styled("div", {
-  shouldForwardProp: shouldForwardStepperDotProps,
-})<{
-  status: "previous" | "current" | "next";
-  odysseyDesignTokens: DesignTokens;
-  isClickable: boolean;
-}>(({ status, odysseyDesignTokens, isClickable }) => ({
-  width: odysseyDesignTokens.Spacing2,
-  height: odysseyDesignTokens.Spacing2,
-  borderRadius: "50%",
-  border: "1px solid",
-  borderColor:
-    status === "current"
-      ? odysseyDesignTokens.HueNeutral500
-      : status === "previous"
-        ? odysseyDesignTokens.HueNeutral400
-        : odysseyDesignTokens.HueNeutral300,
-  background:
-    status === "current" ? odysseyDesignTokens.HueNeutral500 : "transparent",
-  margin: "0 2px",
-  cursor: isClickable ? "pointer" : "default",
-  "&:hover": isClickable
-    ? {
-        background: odysseyDesignTokens.HueNeutral300,
-        borderColor: odysseyDesignTokens.HueNeutral500,
-      }
-    : undefined,
-  "&:focus-visible": isClickable
-    ? {
-        boxShadow: `0 0 0 2px ${odysseyDesignTokens.HueNeutralWhite}, 0 0 0 4px ${odysseyDesignTokens.PalettePrimaryMain}`,
-        outline: "2px solid transparent",
-        outlineOffset: "1px",
-      }
-    : undefined,
-}));
-
-const StepperNavigationContainer = styled("div", {
-  shouldForwardProp: shouldForwardStepperNavigationProps,
-})<{
-  odysseyDesignTokens: DesignTokens;
-}>(({ odysseyDesignTokens }) => ({
-  display: "grid",
-  gridTemplateColumns: "1fr auto 1fr",
-  alignItems: "center",
-  marginTop: odysseyDesignTokens.Spacing3,
-  gap: odysseyDesignTokens.Spacing3,
-}));
-
-const NavigationSection = styled("div", {
-  shouldForwardProp: shouldForwardNavigationSectionProps,
-})<{
-  align: "start" | "center" | "end";
-}>(({ align }) => ({
-  display: "flex",
-  justifySelf:
-    align === "start" ? "flex-start" : align === "end" ? "flex-end" : "center",
-  alignItems: "center",
-}));
-
-export type StepperNavigationProps = {
-  /**
-   * Total number of steps
-   */
-  totalSteps: number;
-  /**
-   * Current active step (0-based index)
-   */
-  currentStep: number;
-  /**
-   * Callback fired when back button is clicked
-   */
-  onBack: () => void;
-  /**
-   * Callback fired when next button is clicked
-   */
-  onNext: () => void;
-  /**
-   * Callback fired when a step dot is clicked
-   */
-  onStepClick: (step: number) => void;
-  /**
-   * Function to determine if a step is clickable
-   */
-  isStepClickable: (step: number) => boolean;
-  /**
-   * Custom label for the previous button
-   */
-  previousButtonLabel?: string;
-  /**
-   * Custom label for the next button
-   */
-  nextButtonLabel?: string;
-  /**
-   * Design tokens for styling
-   */
-  odysseyDesignTokens: DesignTokens;
-};
-
-const StepperNavigation = ({
-  totalSteps,
-  currentStep,
-  onBack,
-  onNext,
-  previousButtonLabel,
-  nextButtonLabel,
-  odysseyDesignTokens,
-  onStepClick,
-  isStepClickable,
-}: StepperNavigationProps) => {
-  const { t } = useTranslation();
-
-  const labels = useMemo(
-    () => ({
-      //TODO: Determine if translation is needed or if we can re-use
-      previous: previousButtonLabel ?? t("pagination.previous"),
-      next: nextButtonLabel ?? t("pagination.next"),
-    }),
-    [previousButtonLabel, nextButtonLabel, t],
-  );
-
-  const dots = useMemo(() => {
-    return Array.from({ length: totalSteps }, (_, i) => {
-      const status: "previous" | "current" | "next" =
-        i === currentStep ? "current" : i < currentStep ? "previous" : "next";
-      const isClickable = isStepClickable(i);
-
-      const handleKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-        if (!isClickable) return;
-        // Only handle Enter/Space for selection
-        if (event.key === "Enter" || event.key === " ") {
-          event.preventDefault(); // Prevent page scroll on space
-          onStepClick(i);
-        }
-      };
-
-      return (
-        <StyledStepperDot
-          key={i}
-          status={status}
-          odysseyDesignTokens={odysseyDesignTokens}
-          isClickable={isClickable}
-          onClick={isClickable ? () => onStepClick(i) : undefined}
-          onKeyDown={handleKeyDown}
-          role="button"
-          tabIndex={isClickable ? 0 : -1}
-          aria-label={`Go to step ${i + 1}`} //TODO: Translation string
-        />
-      );
-    });
-  }, [
-    totalSteps,
-    currentStep,
-    isStepClickable,
-    onStepClick,
-    odysseyDesignTokens,
-  ]);
-
-  return (
-    <StepperNavigationContainer odysseyDesignTokens={odysseyDesignTokens}>
-      <NavigationSection align="start">
-        {currentStep > 0 && (
-          <Button
-            label={labels.previous}
-            variant="secondary"
-            onClick={onBack}
-            size="small"
-          />
-        )}
-      </NavigationSection>
-      <NavigationSection align="center">{dots}</NavigationSection>
-      <NavigationSection align="end">
-        {currentStep < totalSteps - 1 && (
-          <Button
-            label={labels.next}
-            variant="secondary"
-            onClick={onNext}
-            size="small"
-          />
-        )}
-      </NavigationSection>
-    </StepperNavigationContainer>
-  );
-};
-const StepIcon = ({
-  completed,
-  active,
-  stepNumber,
-  variant,
-  odysseyDesignTokens,
-  nonLinear,
-}: {
-  completed: boolean;
-  active: boolean;
-  stepNumber: number;
-  variant: "numeric" | "nonNumeric";
-  odysseyDesignTokens: DesignTokens;
-  nonLinear: boolean;
-}) => (
-  <StyledStepIconContainer
-    completed={completed}
-    active={active}
-    variant={variant}
-    odysseyDesignTokens={odysseyDesignTokens}
-    nonLinear={nonLinear}
-  >
-    {completed && variant === "numeric" ? (
-      <CheckIcon />
-    ) : variant === "numeric" ? (
-      <StyledStepNumber
-        completed={completed}
-        active={active}
-        odysseyDesignTokens={odysseyDesignTokens}
-        nonLinear={nonLinear}
-      >
-        {stepNumber + 1}
-      </StyledStepNumber>
-    ) : null}
-  </StyledStepIconContainer>
-);
-
 const Stepper = ({
   activeStep,
   allowBackStep = false,
@@ -646,12 +328,56 @@ const Stepper = ({
 }: StepperProps) => {
   const odysseyDesignTokens = useOdysseyDesignTokens();
 
+  const StepIcon = ({
+    completed,
+    active,
+    stepNumber,
+    variant,
+    odysseyDesignTokens,
+    nonLinear,
+  }: {
+    completed: boolean;
+    active: boolean;
+    stepNumber: number;
+    variant: "numeric" | "nonNumeric";
+    odysseyDesignTokens: DesignTokens;
+    nonLinear: boolean;
+  }) => (
+    <StyledStepIconContainer
+      completed={completed}
+      active={active}
+      variant={variant}
+      odysseyDesignTokens={odysseyDesignTokens}
+      nonLinear={nonLinear}
+    >
+      {completed && variant === "numeric" ? (
+        <CheckIcon />
+      ) : variant === "numeric" ? (
+        <StyledStepNumber
+          completed={completed}
+          active={active}
+          odysseyDesignTokens={odysseyDesignTokens}
+          nonLinear={nonLinear}
+        >
+          {stepNumber + 1}
+        </StyledStepNumber>
+      ) : null}
+    </StyledStepIconContainer>
+  );
+
   const isStepClickable = useCallback(
     (stepIndex: number) => {
       const isCompleted = stepIndex < activeStep;
-      return nonLinear
-        ? (isCompleted && allowBackStep) || !isCompleted
-        : stepIndex === activeStep;
+
+      if (isCompleted && allowBackStep) {
+        return true; // Allow clicking completed steps if allowBackStep is true
+      }
+
+      if (nonLinear) {
+        return !isCompleted; // Allow clicking future steps if nonLinear
+      }
+
+      return stepIndex === activeStep; // Only allow clicking current step in linear mode
     },
     [activeStep, allowBackStep, nonLinear],
   );
@@ -725,9 +451,7 @@ const Stepper = ({
           onKeyDown={(e) => handleKeyDown(e, index)}
           odysseyDesignTokens={odysseyDesignTokens}
           orientation={orientation}
-          isClickable={
-            nonLinear ? (completed && allowBackStep) || !completed : false
-          }
+          isClickable={isStepClickable(index)}
           role="tab"
           tabIndex={isClickable ? 0 : -1}
           {...ariaProps}
@@ -806,16 +530,7 @@ const Stepper = ({
   );
 };
 
-const MemoizedStepperNavigation = memo(StepperNavigation);
-MemoizedStepperNavigation.displayName = "StepperNavigation";
-
-const MemoizedStepIcon = memo(StepIcon);
-MemoizedStepIcon.displayName = "StepIcon";
-
 const MemoizedStepper = memo(Stepper);
 MemoizedStepper.displayName = "Stepper";
 
-export {
-  MemoizedStepper as Stepper,
-  MemoizedStepperNavigation as StepperNavigation,
-};
+export { MemoizedStepper as Stepper };
